+++ backend/scrapers/README.md
@@ -4,7 +4,7 @@ This module extracts raw data from Congress endpoints/web pages and stores it in
 
 ## Files
 
-- `bills.py`, `motions.py`: scrape bill/motion JSON payloads into `RawBill`/`RawMotion`.
+- `bills.py`, `motions.py`, `leyes.py`: scrape bill/motion/leyes JSON/XML payloads into `RawBill`/`RawMotion`/`RawLeyes`.
 - `bills_documents.py`, `motions_documents.py`: extract and OCR documents into raw document tables.
 - `congresistas.py`, `bancadas.py`, `committees.py`, `organizations.py`: scrape reference entities.
 - `utils.py`: HTTP helpers, parsing helpers, OCR/PDF support.
